Safari-609.1.11.1/LayoutTests/svg/dom/SVGAnimatedLengthList-expected.txt

This test checks the SVGAnimatedLengthList API - utilizing the dx property of SVGTextElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial dx value
PASS textElement.dx.toString() is "[object SVGAnimatedLengthList]"
PASS textElement.dx.baseVal.toString() is "[object SVGLengthList]"
PASS textElement.dx.baseVal.getItem(0).value is 50

Check that length lists are dynamic, caching value in a local variable and modifying it, should take effect
PASS numRef.getItem(0).value is 100
PASS textElement.dx.baseVal.getItem(0).value is 100

Check that assigning to baseVal has no effect, as no setter is defined
PASS textElement.dx.baseVal = -1 is -1
PASS textElement.dx.baseVal = 'aString' is "aString"
PASS textElement.dx.baseVal = textElement is textElement

Check that the dx value remained 100, and the baseVal type has not been changed
PASS textElement.dx.baseVal.toString() is "[object SVGLengthList]"
PASS textElement.dx.baseVal.getItem(0).value is 100
PASS successfullyParsed is true

TEST COMPLETE
